refactor to use seperate javascript files for all routes
