+++ frontend/index.html
@@ -8,8 +8,6 @@
     <link rel="stylesheet" href="/css/base.css">
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
     <script src="/js/app.js" type="module"></script>
-    <script src="/js/api.js" type="module"></script>
-    <script src="/js/loginHandlers.js" type="module"></script>
     <link href="https://fonts.googleapis.com/css?family=Glegoo" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css?family=Inter" rel="stylesheet">
 </head>
