+++ frontend/routes/404.html
@@ -1,3 +1,1 @@
-<div id="content">
-    <h1>404 Page not found</h1>
-</div>
+<h1>404 Page not found</h1>
