+++ frontend/routes/game_menu.html
@@ -1,9 +1,7 @@
-<div id="content">
-    <nav class="menu">
-        <ol>
-            <li><a href="/game_local" data-router-link>local</a></li>
-            <li><a href="/game_online" data-router-link>online</a></li>
-            <li><a href="/main_menu" data-router-link>back</a></li>
-        </ol>
-    </nav>
-</div>
+<nav class="menu">
+    <ol>
+        <li><a href="/game_local" data-router-link>local</a></li>
+        <li><a href="/game_online" data-router-link>online</a></li>
+        <li><a href="/main_menu" data-router-link>back</a></li>
+    </ol>
+</nav>
