+++ frontend/routes/home.html
@@ -1,69 +1,42 @@
-<div id="content">
-    <form class="needs-validation" novalidate>
-        <span class="d-flex justify-content-center">
-            <span class="login d-inline-flex justify-content-center flex-column">
-                <h2>Sign in or create account</h2>
-                <label for="email">email</label>
-                <input type="email" id="email" autocomplete="email" size="30" required
-                    placeholder="enter your email address" />
-                <div id="validationServerUsernameFeedback" class="invalid-feedback">
-                    Please enter a valid email
-                </div>
-                <button type="submit" class="mainButton" id="login">Continue with email</button>
-                <div class="middle-line-text">
-                    <span>or use on of these options</span>
-                </div>
-                <div>
-                <button type="button" class="secondaryButton btn btn-primary">
-                    <svg version="1.1" id="Calque_1" sodipodi:docname="42_logo.svg" inkscape:version="0.48.2 r9819"
-                        xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#" xmlns:svg="http://www.w3.org/2000/svg"
-                        xmlns:sodipodi="http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd"
-                        xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:cc="http://creativecommons.org/ns#"
-                        xmlns:inkscape="http://www.inkscape.org/namespaces/inkscape" xmlns="http://www.w3.org/2000/svg"
-                        xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px" viewBox="0 -200 960 960"
-                        enable-background="new 0 -200 960 960" xml:space="preserve"
-                        width="16" height="16" >
-                        <polygon id="polygon5" points="32,412.6 362.1,412.6 362.1,578 526.8,578 526.8,279.1 197.3,279.1 526.8,-51.1 362.1,-51.1 
-                    32,279.1 " />
-                        <polygon id="polygon7" points="597.9,114.2 762.7,-51.1 597.9,-51.1 " />
-                        <polygon id="polygon9"
-                            points="762.7,114.2 597.9,279.1 597.9,443.9 762.7,443.9 762.7,279.1 928,114.2 928,-51.1 762.7,-51.1 " />
-                        <polygon id="polygon11" points="928,279.1 762.7,443.9 928,443.9 " />
-                    </svg>
-                    <path
-                        d="M15.545 6.558a9.4 9.4 0 0 1 .139 1.626c0 2.434-.87 4.492-2.384 5.885h.002C11.978 15.292 10.158 16 8 16A8 8 0 1 1 8 0a7.7 7.7 0 0 1 5.352 2.082l-2.284 2.284A4.35 4.35 0 0 0 8 3.166c-2.087 0-3.86 1.408-4.492 3.304a4.8 4.8 0 0 0 0 3.063h.003c.635 1.893 2.405 3.301 4.492 3.301 1.078 0 2.004-.276 2.722-.764h-.003a3.7 3.7 0 0 0 1.599-2.431H8v-3.08z">
-                    </path>
-                    </svg>
-                    Sign in with 42
-                </button>
-
-                <button type="button" class="secondaryButton btn btn-primary">
-                    <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor"
-                        class="bi bi-google" viewBox="0 0 16 16">
-                        <path
-                            d="M15.545 6.558a9.4 9.4 0 0 1 .139 1.626c0 2.434-.87 4.492-2.384 5.885h.002C11.978 15.292 10.158 16 8 16A8 8 0 1 1 8 0a7.7 7.7 0 0 1 5.352 2.082l-2.284 2.284A4.35 4.35 0 0 0 8 3.166c-2.087 0-3.86 1.408-4.492 3.304a4.8 4.8 0 0 0 0 3.063h.003c.635 1.893 2.405 3.301 4.492 3.301 1.078 0 2.004-.276 2.722-.764h-.003a3.7 3.7 0 0 0 1.599-2.431H8v-3.08z">
-                        </path>
-                    </svg>
-                    Sign in with Google
-                </button>
-            </div>
-            </span>
-        </span>
-    </form>
-</div>
-
-<!-- <div id="content">
+<script src="/js/home.js" type="module"></script>
+<form class="needs-validation" novalidate>
     <span class="d-flex justify-content-center">
         <span class="login d-inline-flex justify-content-center flex-column">
-            <form class="d-inline-flex justify-content-center flex-column">
-                <h2>Sign in or create account</h2>
-                <input type="email" id="email" autocomplete="email" size="30" required  placeholder="enter your email address"/>
-                <button class="mainButton" id="login">Continue with email</button>
-                <div class="middle-line-text">
-                    <span>or use on of these options</span>
-                </div>
-            </form>
-            <button class="mainButton" id="42login" formaction="https://profile.intra.42.fr">42</button>
+            <h2>Sign in or create account</h2>
+            <label for="email">email</label>
+            <input type="email" id="email" autocomplete="email" size="30" required
+                placeholder="enter your email address" />
+            <div id="validationServerUsernameFeedback" class="invalid-feedback">
+                Please enter a valid email
+            </div>
+            <button type="submit" class="mainButton" id="login">Continue with email</button>
+            <div class="middle-line-text">
+                <span>or</span>
+            </div>
+            <div>
+            <a href="/api/42auth/redirect/" type="button" class="secondaryButton btn btn-primary">
+                <svg version="1.1" id="Calque_1" sodipodi:docname="42_logo.svg" inkscape:version="0.48.2 r9819"
+                    xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#" xmlns:svg="http://www.w3.org/2000/svg"
+                    xmlns:sodipodi="http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd"
+                    xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:cc="http://creativecommons.org/ns#"
+                    xmlns:inkscape="http://www.inkscape.org/namespaces/inkscape" xmlns="http://www.w3.org/2000/svg"
+                    xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px" viewBox="0 -200 960 960"
+                    enable-background="new 0 -200 960 960" xml:space="preserve"
+                    width="16" height="16" >
+                    <polygon id="polygon5" points="32,412.6 362.1,412.6 362.1,578 526.8,578 526.8,279.1 197.3,279.1 526.8,-51.1 362.1,-51.1 
+                32,279.1 " />
+                    <polygon id="polygon7" points="597.9,114.2 762.7,-51.1 597.9,-51.1 " />
+                    <polygon id="polygon9"
+                        points="762.7,114.2 597.9,279.1 597.9,443.9 762.7,443.9 762.7,279.1 928,114.2 928,-51.1 762.7,-51.1 " />
+                    <polygon id="polygon11" points="928,279.1 762.7,443.9 928,443.9 " />
+                </svg>
+                <path
+                    d="M15.545 6.558a9.4 9.4 0 0 1 .139 1.626c0 2.434-.87 4.492-2.384 5.885h.002C11.978 15.292 10.158 16 8 16A8 8 0 1 1 8 0a7.7 7.7 0 0 1 5.352 2.082l-2.284 2.284A4.35 4.35 0 0 0 8 3.166c-2.087 0-3.86 1.408-4.492 3.304a4.8 4.8 0 0 0 0 3.063h.003c.635 1.893 2.405 3.301 4.492 3.301 1.078 0 2.004-.276 2.722-.764h-.003a3.7 3.7 0 0 0 1.599-2.431H8v-3.08z">
+                </path>
+                </svg>
+                Sign in with 42
+            </a>
+        </div>
         </span>
     </span>
-</div> -->
+</form>
