+++ frontend/routes/login.html
@@ -1,8 +1,6 @@
-<div id="content">
-    <h1>Enter your password</h1>
-    <form id="login-form">
-        <input type="password" name="password" id="password-field" class="login-form-field" placeholder="Password">
-        <input type="submit" value="Login" id="login-form-submit">
-    </form>
-    <a href="/password_reset" data-router-link>Forgot password?</a>
-</div>
+<h1>Enter your password</h1>
+<form id="login-form">
+    <input type="password" name="password" id="password-field" class="login-form-field" placeholder="Password">
+    <input type="submit" value="Login" id="login-form-submit">
+</form>
+<a href="/password_reset" data-router-link>Forgot password?</a>
