+++ frontend/routes/main_menu.html
@@ -1,9 +1,7 @@
-<div id="content">
-    <nav class="menu">
-        <ul>
-            <li><a href="/game_menu" data-router-link>play</a></li>
-            <li><a href="/tournament" data-router-link>tournament</a></li>
-            <li><a href="/chat" data-router-link>chat</a></li>
-        </ul>
-    </nav>
-</div>
+<nav class="menu">
+    <ul>
+        <li><a href="/game_menu" data-router-link>play</a></li>
+        <li><a href="/tournament" data-router-link>tournament</a></li>
+        <li><a href="/chat" data-router-link>chat</a></li>
+    </ul>
+</nav>
